+++ griptape/drivers/event_listener/base_event_listener_driver.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
 import logging
-import threading
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
 from attrs import Factory, define, field
 
+from griptape.mixins.exponential_backoff_mixin import ExponentialBackoffMixin
 from griptape.mixins.futures_executor_mixin import FuturesExecutorMixin
 
 if TYPE_CHECKING:
@@ -16,10 +16,9 @@ logger = logging.getLogger(__name__)
 
 
 @define
-class BaseEventListenerDriver(FuturesExecutorMixin, ABC):
+class BaseEventListenerDriver(FuturesExecutorMixin, ExponentialBackoffMixin, ABC):
     batched: bool = field(default=True, kw_only=True)
     batch_size: int = field(default=10, kw_only=True)
-    thread_lock: threading.Lock = field(default=Factory(lambda: threading.Lock()))
 
     _batch: list[dict] = field(default=Factory(list), kw_only=True)
 
@@ -27,8 +26,21 @@ class BaseEventListenerDriver(FuturesExecutorMixin, ABC):
     def batch(self) -> list[dict]:
         return self._batch
 
-    def publish_event(self, event: BaseEvent | dict, *, flush: bool = False) -> None:
-        self.futures_executor.submit(self._safe_try_publish_event, event, flush=flush)
+    def publish_event(self, event: BaseEvent | dict) -> None:
+        event_payload = event if isinstance(event, dict) else event.to_dict()
+
+        if self.batched:
+            self._batch.append(event_payload)
+            if len(self.batch) >= self.batch_size:
+                self.futures_executor.submit(self._safe_publish_event_payload_batch, self.batch)
+                self._batch = []
+        else:
+            self.futures_executor.submit(self._safe_publish_event_payload, event_payload)
+
+    def flush_events(self) -> None:
+        if self.batch:
+            self.futures_executor.submit(self._safe_publish_event_payload_batch, self.batch)
+            self._batch = []
 
     @abstractmethod
     def try_publish_event_payload(self, event_payload: dict) -> None: ...
@@ -36,18 +48,16 @@ class BaseEventListenerDriver(FuturesExecutorMixin, ABC):
     @abstractmethod
     def try_publish_event_payload_batch(self, event_payload_batch: list[dict]) -> None: ...
 
-    def _safe_try_publish_event(self, event: BaseEvent | dict, *, flush: bool) -> None:
-        try:
-            event_payload = event if isinstance(event, dict) else event.to_dict()
-
-            if self.batched:
-                with self.thread_lock:
-                    self._batch.append(event_payload)
-                    if len(self.batch) >= self.batch_size or flush:
-                        self.try_publish_event_payload_batch(self.batch)
-                        self._batch = []
-                return
-            else:
+    def _safe_publish_event_payload(self, event_payload: dict) -> None:
+        for attempt in self.retrying():
+            with attempt:
                 self.try_publish_event_payload(event_payload)
-        except Exception as e:
-            logger.error(e)
+        else:
+            logger.error("event listener driver failed after all retry attempts")
+
+    def _safe_publish_event_payload_batch(self, event_payload_batch: list[dict]) -> None:
+        for attempt in self.retrying():
+            with attempt:
+                self.try_publish_event_payload_batch(event_payload_batch)
+        else:
+            logger.error("event listener driver failed after all retry attempts")